requirements.txt
pyyaml
pillow
numpy
# opencv-python
# dearpygui
# screeninfo